+++ packages/core/android/src/main/java/io/sentry/react/RNSentryModuleImpl.java
@@ -934,7 +934,7 @@ public class RNSentryModuleImpl {
   }
 
   public void getNewScreenTimeToDisplay(Promise promise) {
-    RNSentryTimeToDisplay.getTimeToDisplay(promise, dateProvider);
+    RNSentryTimeToDisplay.getTimeToDisplay(promise, dateProvider, logger);
   }
 
   private String getProfilingTracesDirPath() {